// src/generated/trading212.ts
export interface components {
    schemas: {
// ...
        ReportResponse: {
            dataIncluded?: components["schemas"]["ReportDataIncluded"];
            downloadLink?: string;
            /** Format: int64 */
            reportId?: number;
            /** @enum {string} */
            status?: "Queued" | "Processing" | "Running" | "Canceled" | "Failed" | "Finished";
            /** Format: date-time */
            timeFrom?: string;
            /** Format: date-time */
            timeTo?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}